pyproject.toml
[project]
name = "secdashboards"
version = "0.1.0"
description = "Security Data Lake Analytics - Connect to AWS Security Lake, create detection rules, and deploy to Lambda"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.13"
dependencies = [
    "marimo>=0.10.0",
    "boto3>=1.35.0",
    "pandas>=2.2.0",
    "pyarrow>=18.0.0",
    "polars>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "rich>=13.9.0",
    "typer>=0.15.0",
    "jinja2>=3.1.0",
    "croniter>=5.0.0",
    "structlog>=24.4.0",
    "pyvis>=0.3.2",
    "gremlinpython>=3.7.0",
    "networkx>=3.2",
    "plotly>=5.24.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
    "ty>=0.0.1a6",
    "boto3-stubs[essential,athena,s3,lambda,events,glue,securitylake,neptune]>=1.35.0",
    "moto>=5.1.20",
    "pre-commit>=4.5.1",
    "cfn-lint>=1.43.3",
]
cdk = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[project.scripts]
secdash = "secdashboards.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/secdashboards"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PTH"]

[tool.ruff.lint.per-file-ignores]
# Allow longer lines in test files (test data, assertions)
"tests/*.py" = ["E501"]
"infrastructure/cdk/tests/*.py" = ["E501"]
# Marimo notebooks use bare expressions to render UI elements
"notebooks/*.py" = ["B018", "E501"]
# Scripts contain long strings (ARNs, markdown content, demo data)
"scripts/*.py" = ["E501"]
# CDK stacks contain long resource strings and CSP policies
"infrastructure/cdk/stacks/*.py" = ["E501"]

[tool.ty]
# Ty (Astral's type checker) configuration
# Docs: https://docs.astral.sh/ty/

[tool.ty.environment]
python-version = "3.13"

[tool.ty.rules]
# Ignore unresolved imports for optional dependencies
unresolved-import = "warn"
# Allow missing type stubs for third-party libraries
possibly-missing-attribute = "warn"

[tool.ty.ignore]
# Optional dependencies that may not be installed
paths = [
    "src/secdashboards/reports/exporters.py",  # playwright is optional
]
